Products Module Enhancement & Remove Role Product Owner

+++ Trainer/Products.Categories.Core/Interfaces/IProductsCategoriesManager.cs
@@ -10,8 +10,8 @@ namespace Products.Core.Interfaces
     {
         ResultMessage GetAll();
         ResultMessage Insert(ProductsCategoryDto category);
-        ResultMessage GetById(byte id);
-        ResultMessage Update(ProductsCategoryDto category, byte id);
-        ResultMessage Delete(byte id);
+        ResultMessage GetById(int id);
+        ResultMessage Update(ProductsCategoryDto category, int id);
+        ResultMessage Delete(int id);
     }
 }
